requirements.txt
Flask
python-dotenv
appwrite
gunicorn
requests
werkzeug
